Fix #684: Clean up sbt plugin
- Move things we don't want to be imported into *.sbt files
  automatically to a separate object
- Classify keys using KeyRanks
